# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Grocery & Expiry Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ItemStatus = (typeof ItemStatus)[keyof typeof ItemStatus];

export const ItemStatus = {
  active: "active",
  consumed: "consumed",
  wasted: "wasted",
} as const;

export interface Item {
  id: number;
  name: string;
  category: string;
  quantity: number;
  unit: string;
  purchaseDate?: string;
  expiryDate: string;
  status: ItemStatus;
  estimatedCost?: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateItemRequest {
  name: string;
  category: string;
  quantity: number;
  unit: string;
  purchaseDate?: string;
  expiryDate: string;
  estimatedCost?: number;
}

export type UpdateItemRequestStatus =
  (typeof UpdateItemRequestStatus)[keyof typeof UpdateItemRequestStatus];

export const UpdateItemRequestStatus = {
  active: "active",
  consumed: "consumed",
  wasted: "wasted",
} as const;

export interface UpdateItemRequest {
  name?: string;
  category?: string;
  quantity?: number;
  unit?: string;
  expiryDate?: string;
  status?: UpdateItemRequestStatus;
  estimatedCost?: number;
}

export type ScanReceiptResponseExtractedItemsItem = {
  name: string;
  quantity: number;
  unit: string;
  estimatedCost?: number;
};

export interface ScanReceiptResponse {
  receiptId: number;
  extractedItems: ScanReceiptResponseExtractedItemsItem[];
  message: string;
}

export interface GenerateRecipesRequest {
  expiringItems: string[];
}

export interface RecipeSuggestion {
  name: string;
  description: string;
  ingredients: string[];
  cookTime: string;
  difficulty: string;
}

export interface GenerateRecipesResponse {
  recipes: RecipeSuggestion[];
  expiringItems: string[];
}

export interface WasteAnalytics {
  totalConsumed: number;
  totalWasted: number;
  totalActive: number;
  moneySaved: number;
  moneyWasted: number;
  carbonSaved: number;
  wasteAvoidanceRate: number;
  streak: number;
  level: string;
}

export type ListItemsParams = {
  status?: ListItemsStatus;
};

export type ListItemsStatus =
  (typeof ListItemsStatus)[keyof typeof ListItemsStatus];

export const ListItemsStatus = {
  active: "active",
  consumed: "consumed",
  wasted: "wasted",
} as const;

export type ScanReceiptBody = {
  file: Blob;
};