# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Grocery & Expiry Tracker API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all grocery items
 */
export const ListItemsQueryParams = zod.object({
  status: zod.enum(["active", "consumed", "wasted"]).optional(),
});

export const ListItemsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  quantity: zod.number(),
  unit: zod.string(),
  purchaseDate: zod.coerce.date().optional(),
  expiryDate: zod.coerce.date(),
  status: zod.enum(["active", "consumed", "wasted"]),
  estimatedCost: zod.number().optional(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListItemsResponse = zod.array(ListItemsResponseItem);

/**
 * @summary Add a new grocery item
 */
export const CreateItemBody = zod.object({
  name: zod.string(),
  category: zod.string(),
  quantity: zod.number(),
  unit: zod.string(),
  purchaseDate: zod.coerce.date().optional(),
  expiryDate: zod.coerce.date(),
  estimatedCost: zod.number().optional(),
});

/**
 * @summary Get a single item
 */
export const GetItemParams = zod.object({
  id: zod.coerce.number(),
});

export const GetItemResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  quantity: zod.number(),
  unit: zod.string(),
  purchaseDate: zod.coerce.date().optional(),
  expiryDate: zod.coerce.date(),
  status: zod.enum(["active", "consumed", "wasted"]),
  estimatedCost: zod.number().optional(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update item status or details
 */
export const UpdateItemParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateItemBody = zod.object({
  name: zod.string().optional(),
  category: zod.string().optional(),
  quantity: zod.number().optional(),
  unit: zod.string().optional(),
  expiryDate: zod.coerce.date().optional(),
  status: zod.enum(["active", "consumed", "wasted"]).optional(),
  estimatedCost: zod.number().optional(),
});

export const UpdateItemResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  quantity: zod.number(),
  unit: zod.string(),
  purchaseDate: zod.coerce.date().optional(),
  expiryDate: zod.coerce.date(),
  status: zod.enum(["active", "consumed", "wasted"]),
  estimatedCost: zod.number().optional(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete an item
 */
export const DeleteItemParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Upload and scan a receipt image (OCR pipeline)
 */
export const ScanReceiptBody = zod.object({
  file: zod.instanceof(File),
});

export const ScanReceiptResponse = zod.object({
  receiptId: zod.number(),
  extractedItems: zod.array(
    zod.object({
      name: zod.string(),
      quantity: zod.number(),
      unit: zod.string(),
      estimatedCost: zod.number().optional(),
    }),
  ),
  message: zod.string(),
});

/**
 * @summary Generate recipe suggestions from expiring ingredients
 */
export const GenerateRecipesBody = zod.object({
  expiringItems: zod.array(zod.string()),
});

export const GenerateRecipesResponse = zod.object({
  recipes: zod.array(
    zod.object({
      name: zod.string(),
      description: zod.string(),
      ingredients: zod.array(zod.string()),
      cookTime: zod.string(),
      difficulty: zod.string(),
    }),
  ),
  expiringItems: zod.array(zod.string()),
});

/**
 * @summary Get waste avoided analytics and gamification metrics
 */
export const GetWasteAnalyticsResponse = zod.object({
  totalConsumed: zod.number(),
  totalWasted: zod.number(),
  totalActive: zod.number(),
  moneySaved: zod.number(),
  moneyWasted: zod.number(),
  carbonSaved: zod.number(),
  wasteAvoidanceRate: zod.number(),
  streak: zod.number(),
  level: zod.string(),
});