# lib/api-zod/src/generated/types/generateRecipesRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Grocery & Expiry Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateRecipesRequest {
  expiringItems: string[];
}