# lib/api-zod/src/generated/types/generateRecipesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Grocery & Expiry Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { RecipeSuggestion } from "./recipeSuggestion";

export interface GenerateRecipesResponse {
  recipes: RecipeSuggestion[];
  expiringItems: string[];
}