# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Grocery & Expiry Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemStatus } from "./itemStatus";

export interface Item {
  id: number;
  name: string;
  category: string;
  quantity: number;
  unit: string;
  purchaseDate?: Date;
  expiryDate: Date;
  status: ItemStatus;
  estimatedCost?: number;
  createdAt: Date;
  updatedAt: Date;
}