# lib/api-zod/src/generated/types/recipeSuggestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Grocery & Expiry Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface RecipeSuggestion {
  name: string;
  description: string;
  ingredients: string[];
  cookTime: string;
  difficulty: string;
}