# lib/api-zod/src/generated/types/scanReceiptResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Grocery & Expiry Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanReceiptResponseExtractedItemsItem } from "./scanReceiptResponseExtractedItemsItem";

export interface ScanReceiptResponse {
  receiptId: number;
  extractedItems: ScanReceiptResponseExtractedItemsItem[];
  message: string;
}