# lib/api-zod/src/generated/types/updateItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Grocery & Expiry Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateItemRequestStatus } from "./updateItemRequestStatus";

export interface UpdateItemRequest {
  name?: string;
  category?: string;
  quantity?: number;
  unit?: string;
  expiryDate?: Date;
  status?: UpdateItemRequestStatus;
  estimatedCost?: number;
}