# lib/api-zod/src/generated/types/wasteAnalytics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Grocery & Expiry Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface WasteAnalytics {
  totalConsumed: number;
  totalWasted: number;
  totalActive: number;
  moneySaved: number;
  moneyWasted: number;
  carbonSaved: number;
  wasteAvoidanceRate: number;
  streak: number;
  level: string;
}